Text Box (can switch between static and editable)
Summary: A couple problems left:

1) I can't re-edit the same textbox. Double clicking just creates a new
Text Element.

UNFIXED: I'll deal with this in a future commit. Priority.

2) Actually just realized this is worse. We creates a new Text Element but
set overall Action to Typing, which means ALL text elements are now editable
again. I think to fix this, every text object will need its own reference

FIXED: gave every text type a "typing" boolean. Depending on whether it's set
we'll be in Editable Text mode or in Static mode.

3) When I double click, it creates a new Text element but my cursor isn't
on it, so I can't immediately start typing in it. This probably needs to
be handled here, in this class. As soon as it renders, the cursor needs
to focus on it. We might need a "typing" textIdRef

FIXED: used a reference and a useEffect hook. If we are in a "editable" mode
(decided by the "typing boolean" from above) AND the reference changes from its
initial null value to a div reference, we focus on this element.

4) I probably need an onDragStart/onDragEnd event handler for all elements.

UNFIXED: I'll deal with it later

5) width and height aren't being passed from TextAreaInput to Static

UNFIXED: I'll deal with it later

6) I can drag elements even if I'm not using "HAND" tool. This is
probably fine, but could cause problems